Split api helper into two pieces.
